# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API
  description: |-
    This is the API for WASAPhoto homework 1 by 1894389
  version: 0.1.0
...
paths:
...
  /users:
     get:
      tags: ["search"]
      summary: Finds the profile of an user by his Id
      description: Gets an user profile
      operationId: getUserProfile
      parameters:
      - name: id
        in: query
        description: ID dell'utente che viene cercato
        schema:
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
        example: "Luis64"
      responses:
        '200':
          $ref: "#/components/responses/user_ok"
        '400':
          $ref: "#/components/responses/bad_request"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: "#/components/responses/forbidden"
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
# --------------------------------------------------------------
...
  /users/{id}/banned_users/{banned_id}:
    parameters:
        - $ref: "#/components/parameters/banned_id"
        - $ref: '#/components/parameters/identifier'
    put:
      tags: ["ban"]
      summary: Bans/block a user
      description: Once a user is blocked he/she won't be able to see any info from whoever banned him/her(photo,comments,likes,ecc...).
      operationId: banUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: [] 
    
    delete:
      tags: ["ban"]
      summary: Unbans/unblocks a user
      description: By removing the ban,the unbanned user can see every information of the persone that unbanned him/her.
      operationId: unbanUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      
      security:
        - bearerAuth: []
# --------------------------------------------------------------
  /users/{id}/followers/{follower_id}:
    parameters:
        - $ref: "#/components/parameters/follower_id"
        - $ref: '#/components/parameters/identifier'
        
    put:
      tags: ["followers"]
      summary: Follows a specific user
      description: Allows a user to follow another user. 
      operationId: followUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      
      security:
        - bearerAuth: [] 
          
    delete:
      tags: ["followers"]
      summary: Unfollows a specific user
      description: Unfollow another one,removing him from your list of followed users.
      operationId: unfollowUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"      
          
      security:
        - bearerAuth: []
# --------------------------------------------------------------
...
  /users/{id}/photos/{photo_id}:
    parameters: 
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/photo_id'
        
        
    get:
      tags: ["photo"]
      summary: Downloads a photo
      description: A user can access a photo and download it
      operationId: getPhoto
      
      responses:
        '200':
          $ref: "#/components/responses/photo_download"
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: [] 
        
    delete:
      tags: ["photo"]
      summary: Deletes a photo
      description: Deletes a user's photo with relative comments and likes
      operationId: deletePhoto
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: [] 
# --------------------------------------------------------------  
  /users/{id}/photos/{photo_id}/comments:
    parameters: 
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/photo_id'
        
    post:
      tags: ["comments"]
      summary: Add a comment to a photo
      description: Adds a comment to a user's photo.
      operationId: commentPhoto
      
      requestBody:
        description: Contains the id of who's commenting and the relative comment
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/comment'
        required: true
      
      responses:
        '201':
          $ref: '#/components/responses/comment_added'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: [] 
# --------------------------------------------------------------
...
  /users/{id}/photos/{photo_id}/likes/{like_id}:
    parameters: 
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/photo_id'
        - $ref: '#/components/parameters/like_id'
  
    put:
      tags: ["likes"]
      summary: Add a like to a photo
      description: Adds a like to a user's photo. A user can only like a photo once.
      operationId: likePhoto
      
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      
      security:
        - bearerAuth: [] 
    
    delete:
      tags: ["likes"]
      summary: Remove a like from a photo
      description: Removes a like from a user's photo
      operationId: unlikePhoto
      
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: [] 
      
# --------------------------------------------------------------
components:
  parameters:
    identifier:
      name: id
      in: path
      description: It's a user's unique id
      required: true
      schema:
        description: String between 3-16 characters
        type: string
        pattern: '^.*?$'
        example: Micheal0Jackson
        minLength: 3
        maxLength: 16
        readOnly: true
    banned_id:
      name: banned_id
      in: path
      description: A banned user identifier.
      required: true
      schema:
        $ref: "#/components/schemas/user_nickname/properties/nickname"
      example: "PerverseGuy"
    follower_id:
      name: follower_id
      in: path
      description: A follower's identifier.
      required: true
      schema: 
        $ref: "#/components/schemas/user_nickname/properties/nickname"
      example: "Pikachu1"
    photo_id:
      name: photo_id
      in: path
      description: A photo identifier
      required: true
      schema: 
        description: Int64 between 0-9999
        type: integer
        format: int64
        minimum: 0
        maximum: 9999
        example : 999
        readOnly: true
...
    like_id:
      name: like_id
      in: path
      description: A like unique identifier.
      required: true
      schema: 
        $ref: "#/components/schemas/user_nickname/properties/nickname"
      example: "Mami"
      
# --------------------------------------------------------------
  schemas:
...
    full_user:
      description: Contains the *unique identifier and the nickname of a user
      type: object
      properties:
        user_id:
          $ref: "#/components/schemas/user_nickname/properties/nickname"
        nickname:
          $ref: "#/components/schemas/user_nickname/properties/nickname"
      example:
        user_id: "John64"
        nickname: "NeX"
...
    users_list:
      description: List of users found
      type: object
      properties:
        users:
          description: Array of users with a similar/partial name from the searched one
          type: array
          minItems: 0
          maxItems: 10
          items:
            $ref: "#/components/schemas/full_user"
          example:
            - user_id: jonathan
              nickname: jojo1
            - user_id: joseph
              nickname: jojo2
            - user_id: jotaro
              nickname: jojo3
    comment_identifier:
      description: It's a comment id
      type: object
      properties:
        comment_id:
          description: Any number between 0-9999
          type: integer
          format: int64
          minimum: 0
          maximum: 9999
          readOnly: true
          example: 965
      example:
        comment_id: 965
    comment:
      description: A user's comment containing the text of the comment and the id of who commented
      type: object
      properties:
        user_id: 
          $ref: "#/components/schemas/user_nickname/properties/nickname"
        nickname: 
          $ref: "#/components/schemas/user_nickname/properties/nickname"
        comment:
          type: string
          minLength: 1
          maxLength: 30
          pattern: '^.*?$'
          example: Cool brother
          description: String value of the comment
        comment_id:
          $ref: "#/components/schemas/comment_identifier/properties/comment_id"
      example:
        user_id: "Kiddo123"
        nickname: "Bad_Kiddo"
        photo_id: 873
        comment: "WOW WHAT A SCENERY"    
        comment_id: 16
    image:
      description: Image content
      type: object
      properties:
        photo_data:
          description: "Raw data of the image/photo"
          type: string
          minLength: 0
          maxLength: 100
          format: binary
          example: "101010101"
      example: 
        photo_data: "011110100"
...
  responses:
    user_ok:
      description: User found successfully by the id/nick.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/users_list"
...
    photo_download:
      description: This response is associated to the 200 http status(Gets u the image file of the photo u downloaded)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/image"
...
    comment_added:
      description: This response is associated to the 201 http status(Comment has been added successfully.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/comment_identifier"
    no_content:
      description: This response is associated to the 204 http status (There is no content to send for this request)
...
    bad_request:
      description: This response is associated to the 400 http status (The server cannot process the request due to something that is perceived to be a client error)
    unauthorized:
      description: This response is associated to the 401 http status (The client must authenticate itself to get the requested response)
    forbidden:
      description: This response is associated to the 403 http status (The client does not have access rights to the content)
    not_found:
      description: This response is associated to the 404 http status (The server cannot find the requested resource.)
    internal_server_error:
      description: This response is associated to the 500 http status (The server has encountered a situation it does not know how to handle.)

# --------------------------------------------------------------
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT